services/quiz-service/requirements.txt
fastapi
uvicorn
confluent-kafka
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
google-generativeai
pytest
httpx
